requirements.txt
# Core dependencies — for production: pip freeze > requirements-lock.txt recommended
pymupdf>=1.24.0,<2
pandas>=2.0,<3
pytest>=8.0
jupyter>=1.0
sentence-transformers>=3.0,<4
faiss-cpu>=1.8.0
numpy>=1.24,<3
openai>=1.40,<2
python-dotenv>=1.0
fastapi>=0.109,<1
uvicorn[standard]>=0.24
streamlit>=1.28
slowapi>=0.1.9,<0.2
httpx>=0.27
